File creation file dialog prompt
Implemented a file creation function that will prompt the user in file explorer to create a new file and select a folder for it to be created in. This improves user experience so they can choose any location on their PC without the need to type a relative path from the app directory.